Passing the $element as context to itemSelected
With the current implementation it is not possible to access this.$element
from the itemSelected function. This comes handy when implementing
several typeaheads in the same page and you need to react in other elements
once the typeahead was selected. By using proxy we are passing the current
element in the context and it can be accessed as 'this' from within 
itemSelected. 

Example:
[...]
itemSelected: function (item) {
    $(this).prev('input[type="hidden"]').val(item);
},
[...]